feat(notewave): add converting transcript (if it is array of segments) into contextTextOrigin
